fix(core): dtype promotion bug in trf, no-op padding, and None guard in minpack
- trf.py: remove jnp.float64(0.995) that forces dtype promotion in float32 context
- covariance_computer.py: skip identity padding when len_diff == 0 (was wasteful no-op)
- minpack.py: guard n_starts > 0 against None to prevent TypeError

+++ nlsq/core/trf.py
@@ -2522,7 +2522,7 @@ class TrustRegionReflective(TrustRegionJITFunctions, TrustRegionOptimizerBase):
             )
 
             # theta controls step back step ratio from the bounds
-            theta = jnp.maximum(jnp.float64(0.995), 1.0 - g_norm)
+            theta = jnp.maximum(0.995, 1.0 - g_norm)
 
             # Evaluate inner loop using helper
             inner_result = self._evaluate_bounds_inner_loop(
